Add error callouts and other improvements for Node Communications
This code adds bus, hw, code, etc callouts to the appropriate errors.
It also cleans up other TODOs, refactors some functions, and applies
somes fixes found in testing.

RTC:184518
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/61813
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/usr/secureboot/node_comm/node_comm_test.C
@@ -67,8 +67,8 @@ errlHndl_t nodeCommXbus2ProcTest(void)
     uint64_t read_data = 0;
     Target* read_tgt = nullptr;
     node_comm_modes_t mode = NCDD_MODE_XBUS;
-    uint64_t linkId = 0;
-    uint64_t mboxId = 0;
+    uint8_t linkId = 0;
+    uint8_t mboxId = 0;
     bool attn_found = false;
 
     do
